Extract base Configurator from Configurators namespace
