--- sig/kangaru/configurators/configurator.rbs
@@ -1,14 +1,0 @@
-module Kangaru
-  module Configurators
-    class Configurator
-      include Concerns::AttributesConcern
-      extend  Concerns::AttributesConcern::ClassMethods
-
-      def self.key: -> Symbol
-
-      def self.name: -> String
-
-      def serialise: -> Hash[Symbol, untyped]
-    end
-  end
-end
